+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Changed:
   * sh wrapper: set `working_dir` always, default to `$PWD`
   * Drop all caching, #132, #143, #126, #100, #92, #45, #85
 
+Added:
+
+  * CLI: `--force` option for `workspace add` to replace existing ID, #134
+
 ## [0.4.4] - 2018-07-03
 
 Fixes: