+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.5.0] - 2018-07-19
+
 Fixed:
 
   * EXIF supports JFIF metadata, #141
@@ -203,6 +205,7 @@ Fixed
 Initial Release
 
 <!-- link-labels -->
+[0.5.0]: ../../compare/v0.5.0...v0.4.4
 [0.4.4]: ../../compare/v0.4.4...v0.4.3
 [0.4.3]: ../../compare/v0.4.3...v0.4.2
 [0.4.2]: ../../compare/v0.4.2...v0.4.1